fouten uithalen

+++ RestaurantReservatie.BL/Managers/CustomerManager.cs
@@ -1,29 +1,72 @@
-﻿using RestaurantReservatie.BL.Interfaces;
+﻿using RestaurantReservatie.BL.Exceptions;
+using RestaurantReservatie.BL.Interfaces;
 using RestaurantReservatie.BL.Models;
 
-namespace RestaurantReservatie.BL.Managers; 
+namespace RestaurantReservatie.BL.Managers;
 
-public class CustomerManager  {
-    
+public class CustomerManager {
     ICustomerRepository _customerRepository;
-    
-    public CustomerManager(ICustomerRepository customerRepository){
+
+    public CustomerManager(ICustomerRepository customerRepository) {
         _customerRepository = customerRepository;
     }
-    
-    public void AddCustomer(Customer customer){
-        _customerRepository.AddCustomer(customer);
+
+    public Customer AddCustomer(Customer customer) {
+        try {
+            if (customer == null) throw new CustomerManagerException("AddCustomer - Gebruiker mag niet null zijn");
+            if (_customerRepository.CustomerExists(customer.CustomerId))
+                throw new CustomerManagerException("AddCustomer - Gebruiker bestaat al");
+            return _customerRepository.AddCustomer(customer);
+        }
+        catch (Exception ex) {
+            throw new CustomerManagerException("AddCustomer - Er is een fout opgetreden", ex);
+        }
+    }
+
+    public Customer GetCustomer(int id) {
+        return _customerRepository.GetCustomer( id);
     }
     
-    public List<Customer> GetCustomers(string filter){
-        return _customerRepository.GetCustomers(filter);
+    public List<Customer> GetAllCustomers() {
+        try {
+            return _customerRepository.GetAllCustomers();
+        }
+        catch (Exception ex) {
+            throw new CustomerManagerException("GetAllCustomers - Er is een fout opgetreden", ex);
+        }
     }
-    
-    public void DeleteCustomer(int id){
-        _customerRepository.DeleteCustomer(id);
-    }
-    
-    public void UpdateCustomer(Customer customer){
-        _customerRepository.UpdateCustomer(customer);
+
+    public void DeleteCustomer(int id) {
+        try
+        {
+            if (id == null) throw new CustomerManagerException("DeleteCustomer - Gebruiker mag niet null zijn");
+            if (!_customerRepository.CustomerExists(id)) throw new CustomerManagerException("VerwijderGebruiker - Gebruiker bestaat niet");
+            _customerRepository.DeleteCustomer(id);
+        }
+        catch (Exception ex)
+        {
+            throw new CustomerManagerException("DeleteCustomer - Er is een fout opgetreden", ex);
+        }    }
+
+    public Customer UpdateCustomer(Customer customer) {
+        try
+        {
+            if (customer == null) throw new CustomerManagerException("UpdateGebruiker - Gebruiker mag niet null zijn");
+            if (!_customerRepository.CustomerExists(customer.CustomerId)) throw new CustomerManagerException("UpdateGebruiker - Gebruiker bestaat niet");
+            return _customerRepository.UpdateCustomer(customer);
+        }
+        catch (Exception ex)
+        {
+            throw new CustomerManagerException("UpdateCustomer - Er is een fout opgetreden", ex);
+        }    }
+
+    public bool CustomerExists(int id) {
+        try {
+            if (id < 0) throw new CustomerManagerException("CustomerExists - Id mag niet kleiner dan 0 zijn");
+            return _customerRepository.CustomerExists(id);
+        }
+        catch (Exception ex) {
+            throw new CustomerManagerException("CustomerExists - Er is een fout opgetreden", ex);
+        }
     }
 }
